+++ lib/create.js
@@ -225,3 +225,19 @@ module.exports = (client, req, callback) => {
             break; 
     }
 };
+
+exports.createNodeOfType = (type) => {
+
+    const node = UID();
+
+    type = type.charAt(0).toUpperCase() + type.slice(1);
+
+    return {
+        uid: node,
+        triples: [{
+            subject: node,
+            predicate: RDF_SYNTAX + 'type>',
+            object: FLOW_VOCAB + type + '>'
+        }]
+    };
+};
