+++ lib/stores/cayley/index.js
@@ -33,7 +33,14 @@ module.exports = (config) => {
         .Tag('subject')
         .Out(FLOW_VOCAB + 'json>', 'predicate')
         .Out(RDF_SYNTAX + 'string>')
-        .All();
+        .GetLimit(1);
+    };
+    const getName = (id) => {
+        return client.g.V(id)
+        .Tag('subject')
+        .Out(FLOW_VOCAB + 'name>', 'predicate')
+        .Out(RDF_SYNTAX + 'string>')
+        .GetLimit(1);
     };
     const write = (triples) => {
         if (triples && triples.length) {
@@ -69,6 +76,7 @@ module.exports = (config) => {
     // export store adapter
     return adapters[config] = Object.freeze({
         sequence: (sequence_id, role) => {
+            console.log('Store.sequence:', sequence_id, role);
             return combineAndParse(sequence(client, sequence_id, role));
         },
         entrypoint: (entrypoint_name) => {
@@ -82,6 +90,9 @@ module.exports = (config) => {
             return combineAndParse(visualization.networks(client, id));
         },
         getObject: (id) => {
+            return combineAndParse(getObject(id));
+        },
+        getName: (id) => {
             return combineAndParse(getObject(id));
         },
         incomming: incomming,
