+++ lib/update.js
@@ -1,0 +1,11 @@
+"use strict"
+
+exports.data = (store, node, data) => {
+    // create/update data on node
+};
+exports.name = (store, node, name) => {
+    // create/update name on node
+};
+exports.out = (store, set, out, node) => {
+    // set node delete previous
+};
